seperate `kick_selector` and `kick_target_provider`
