+++ crates/control/src/lib.rs
@@ -15,6 +15,7 @@ pub mod game_controller_state_filter;
 pub mod ground_contact_detector;
 pub mod ground_provider;
 pub mod kick_selector;
+pub mod kick_target_provider;
 pub mod kinematics_provider;
 pub mod led_status;
 pub mod localization;
