+++ crates/hulk_manifest/src/lib.rs
@@ -55,6 +55,7 @@ pub fn collect_hulk_cyclers() -> Result<Cyclers, Error> {
                     "control::ground_contact_detector",
                     "control::ground_provider",
                     "control::kick_selector",
+                    "control::kick_target_provider",
                     "control::kinematics_provider",
                     "control::led_status",
                     "control::localization",
