+++ etc/parameters/default.json
@@ -1073,10 +1073,13 @@
   },
   "kick_selector": {
     "angle_distance_weight": 0.02,
-    "max_kick_around_obstacle_angle": 0.8,
     "kick_pose_obstacle_radius": 0.1,
-    "ball_radius_for_kick_target_selection": 0.15,
     "closer_threshold": 1.0,
+    "goal_accuracy_margin": 0.25,
+    "default_kick_strength": 1.0
+  },
+  "kick_target_provider": {
+    "ball_radius_for_kick_target_selection": 0.15,
     "find_kick_targets": {
       "distance_from_corner": 1.5,
       "corner_kick_target_distance_to_goal": 1.3,
@@ -1084,8 +1087,7 @@
       "max_kick_around_obstacle_angle": 1.0,
       "ball_radius_for_kick_target_selection": 0.15
     },
-    "goal_accuracy_margin": 0.25,
-    "default_kick_strength": 1.0,
+    "max_kick_around_obstacle_angle": 0.8,
     "corner_kick_strength": 0.25
   },
   "behavior": {
